# Make optimizations to native entrypoint for Hermes