--- a/packages/react-strict-dom/src/native/stylex/typed-om/CSSVariableReferenceValue.js
+++ b/packages/react-strict-dom/src/native/stylex/typed-om/CSSVariableReferenceValue.js
@@ -12,39 +12,39 @@ import type { CSSUnparsedValue } from './CSSUnparsedValue';
 // https://drafts.css-houdini.org/css-typed-om-1/#cssvariablereferencevalue
 export class CSSVariableReferenceValue /*extends CSSStyleValue*/ {
   // https://drafts.css-houdini.org/css-typed-om-1/#custom-property-name-string
-  static #validateVariableName(variable: string): void {
+  static _validateVariableName(variable: string): void {
     if (!variable.startsWith('--')) {
       throw new TypeError(`Invalid custom property name: ${variable}`);
     }
   }
 
-  #variable: string;
-  #fallback: CSSUnparsedValue | null;
+  _variable: string;
+  _fallback: CSSUnparsedValue | null;
 
   constructor(variable: string, fallback?: CSSUnparsedValue) {
     if (__DEV__) {
-      CSSVariableReferenceValue.#validateVariableName(variable);
+      CSSVariableReferenceValue._validateVariableName(variable);
     }
     // No super() call because it's slow in Hermes
-    this.#variable = variable;
-    this.#fallback = fallback ?? null;
+    this._variable = variable;
+    this._fallback = fallback ?? null;
   }
 
   get variable(): string {
-    return this.#variable;
+    return this._variable;
   }
 
   set variable(variable: string): void {
-    this.#variable = variable;
+    this._variable = variable;
   }
 
   get fallback(): CSSUnparsedValue | null {
-    return this.#fallback;
+    return this._fallback;
   }
 
   toString(): string {
-    return `var(${this.#variable}${
-      this.#fallback ? `, ${this.#fallback.toString()}` : ''
+    return `var(${this._variable}${
+      this._fallback ? `, ${this._fallback.toString()}` : ''
     })`;
   }
 }