# Commits

* Avoid calling `super()` on native.
* Avoid private class fields.

# Performance comparison

This change was tested e2e using React Native Fantom. The test case was rendering 100 `html.div` elements with various props and the "complex" styles fixture used in this repo's [JS benchmarks](https://github.com/facebook/react-strict-dom/blob/main/packages/benchmarks/perf/tests/css-props-tests.js). Results as follows...

| Test Scenario | Metric | Main | Optimized | % Change |
|---------------|--------|------|-----------|----------|
| **Nested divs** | Latency Avg (ns) | 29,839,920 | 28,962,893 | -2.94% |
| | Latency Median (ns) | 29,454,631 | 28,908,837 | -1.85% |
| | Throughput Avg (ops/s) | 34 | 35 | +2.94% |
| | Throughput Median (ops/s) | 34 | 35 | +2.94% |
| **Sibling divs** | Latency Avg (ns) | 29,703,416 | 29,048,249 | -2.21% |
| | Latency Median (ns) | 29,574,136 | 28,955,402 | -2.09% |
| | Throughput Avg (ops/s) | 34 | 34 | 0% |
| | Throughput Median (ops/s) | 34 | 35 | +2.94% |